Extract tests for backtest_reports to their own test module
